requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
aiofiles
python-dotenv
# sam2 #The segmentation model itself
# torch #PyTorch for SAM2 model